Small visit dto builder adjustments (#974)
* Move into own file

* Use constructor to map entity to dto

* Updated test to check client name
