+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/dto/VisitDto.kt
@@ -63,9 +63,3 @@ data class VisitDto(
   @Schema(description = "External system details associated with the visit")
   val visitExternalSystemDetails: VisitExternalSystemDetailsDto?,
 )
-data class VisitExternalSystemDetailsDto(
-  @Schema(description = "Client name", example = "client_name")
-  val clientName: String?,
-  @Schema(description = "Client visit reference", example = "Reference ID in the client system")
-  val clientVisitReference: String?,
-)
