+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/dto/VisitExternalSystemDetailsDto.kt
@@ -1,0 +1,17 @@
+package uk.gov.justice.digital.hmpps.visitscheduler.dto
+
+import io.swagger.v3.oas.annotations.media.Schema
+import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.VisitExternalSystemDetails
+import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.VisitNote
+
+data class VisitExternalSystemDetailsDto(
+  @Schema(description = "Client name", example = "client_name")
+  val clientName: String?,
+  @Schema(description = "Client visit reference", example = "Reference ID in the client system")
+  val clientVisitReference: String?,
+) {
+  constructor(visitExternalSystemDetailsEntity: VisitExternalSystemDetails) : this (
+    clientName = visitExternalSystemDetailsEntity.clientName,
+    clientVisitReference = visitExternalSystemDetailsEntity.clientReference
+  )
+}
